Subnautica: add Radiation option and adjust location logic
Add a Radiation Rule to the Subnautica options and adapt the location access
logic.
The new Radiation Rule makes it possible to consider zones inside the Radiation
zone of the Aurora without a Radiation Suit, depending on the rule chosen. By
default, players still need the Radiation Suit.

Most locations on the outer circle of the radiation zone are still easily
accessible without the radiation suit (players may need to have fins, a
seaglide, or to heal). This adds a bit more of challenge for those that want it.

Note: there is another rule in place to prevent locations farther than 800 units
from the origin to be considered without a Seaglide. In practice, this removes
most of those locations from the pool of accessible locations if the player
doesn't have a Seaglide (which should avoid making this too tedious).
Only the Grassy Plateaus East Wreck, Lifepod 4 and Lifepod 6 can be considered
by AP without a Seaglide (and they are definitely accessible without one).